fix: split docs section E2E into route-specific files
