+++ e2e/docs/test_documents.py
@@ -5,6 +5,15 @@ from e2e.docs.conftest import _wait_for_pyscript_init
 
 
 @pytest.mark.e2e
+def test_documents_index_shows_section_cards(docs_page_on, assert_no_console_errors):
+    page = docs_page_on("/documents")
+    expect(page.get_by_role("heading", name="Documentation")).to_be_visible()
+    expect(page.get_by_role("heading", name="Getting Started")).to_be_visible()
+    expect(page.get_by_role("link", name="Open Installation")).to_be_visible()
+    expect(page.get_by_role("link", name="Open Signal Stream")).to_be_visible()
+
+
+@pytest.mark.e2e
 def test_documents_page_content(docs_page_on, assert_no_console_errors):
     page = docs_page_on("/documents")
     expect(page.get_by_role("heading", name="Getting Started")).to_be_visible()
